Source_code/requirements.txt
Pillow
numpy
pandas
openpyxl
opencv-python
tk